project/requirements.txt
geopy
folium 
Django>=4.2
djangorestframework>=3.15